pyproject.toml
[tool.poetry]
name = "vaga-aec"
version = "0.1.0"
description = ""
authors = ["berrytern <berrytern@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
uvicorn = {extras = ["standard"], version = "^0.32.1"}
gunicorn = "^23.0.0"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
pydantic = {extras = ["email"], version = "^2.10.2"}
asyncpg = "^0.30.0"
prometheus-client = "^0.21.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"


[tool.pytest.ini_options]
pythonpath = [
    "."
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
